🚚 Move an entity to its project
