--- Entities/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Entities
-{
-    public class Class1
-    {
-
-    }
-}
