+++ ScheduleMasterServer/Controllers/WeatherForecastController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using Entities;
 
 namespace ScheduleMasterServer.Controllers
 {
